Add custom Gazebo plugin
This plug in is a modified version of the original ros_control plugin and that plugin's "default hardware plugin" (confusing, I know)
The modified model plugin provides a /runstop service to match the real robot's driver, and hooks this up to ros_controls existing estop implementation
The modified hardware plugin provides the mode switching services that the real robot has (mocked for now) and publishes joint states that include the virtual joints used by the real platform. The wrist extension joint behaves the same as the driver implementation. The rotation and translation joints aren't hooked up yet, and it may be difficult to provide their functionality without forking the actual "follow_joint_trajectory" implementation, which expects only joints that are present in the URDF
Mirror the real platform setup and use a joint state publisher node to republish the subset of joints that match the URDF