Fix data shape for MaxHeightImage construction
ros_numpy will set the shape of the points buffer from a PCL message to the image dimensions (W,H), but this causes the underlying numba implementation to blow up when it tries to iterate over the points as though they were in a flat array. This change ensures that a flat view of the data is passed through. It does not appear that this extra shape information is used for anything anyway