Tweak move_base config
Use a footprint polygon. The previous default circle was centered on the laser, and thus a particularly poor approximation of the robot's base. This is why the inflation had to be so high
Don't sample y controls (not an omni base)
Lower goal tolerance
Raise minimum theta velocity so the controller has an easier time spinning the robot towards goals directly behind the robot. Up the oscillation timeout to allow enough time for the robot to turn in place without getting flagged as oscillation